refactor: move filesystem watcher tests into update/tests/filesystem_watcher.rs
Follows the parent.rs + parent/ module convention; no behavioral
changes.
